Select for perfect hits in the array data
perfect hits have an e value of 0. I used that to
select the contigs to analyse this time.

Also realised that the de_duplicate data.frame was
actually unnecessary the subset of mito_contigs
would also have the correct number of names to allow
the me to add the names to the contigs.